+++ assets/javascripts/kitten/components/buttons/button/button.stories.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { text, boolean, number, radios, select } from '@storybook/addon-knobs'
-import { Button } from './button'
+import { Button } from './index'
 import { Marger } from '../../layout/marger'
 import { Container } from '../../grid/container'
 import { Grid, GridCol } from '../../grid/grid'
@@ -41,19 +41,36 @@ const modifierOptions = {
   Oxygen: 'oxygen',
   Checked: 'checked',
   Copper: 'copper',
+  Social_facebook: 'social_facebook',
+  Social_twitter: 'social_twitter',
+  Social_linkedin: 'social_linkedin',
+  Social_instagram: 'social_instagram',
+  Social_youtube: 'social_youtube',
+  Social_pinterest: 'social_pinterest',
+}
+
+const sizeOptions = {
+  Nano: 'nano',
+  Micro: 'micro',
+  Tiny: 'tiny',
+  Big: 'big',
+  Huge: 'huge',
+  Giant: 'giant',
+  Normal: 'normal',
 }
 
 export const WithText = () => (
   <Button
-    tiny={boolean('Tiny', false)}
-    big={boolean('Big', false)}
-    huge={boolean('Huge', false)}
-    giant={boolean('Giant', false)}
+    size={select('Size', sizeOptions, 'normal')}
     fluid={boolean('Fluid', false)}
     modifier={select('Modifier', modifierOptions, 'hydrogen')}
     variant={select('Variant', variantOptions, 'andromeda')}
     disabled={boolean('Disabled', false)}
     borderRadius={number('Border radius', 0, radiusBorderRange)}
+    tiny={boolean('Tiny', false)}
+    big={boolean('Big', false)}
+    huge={boolean('Huge', false)}
+    giant={boolean('Giant', false)}
   >
     {text('Text', 'MyButton')}
   </Button>
@@ -75,19 +92,20 @@ WithText.decorators = [
 
 export const WithIcon = () => (
   <Button
+    size={select('Size', sizeOptions, 'normal')}
     rounded={boolean('Rounded', false)}
-    nano={boolean('Nano', false)}
-    micro={boolean('Micro', false)}
-    tiny={boolean('Tiny', false)}
-    big={boolean('Big', false)}
-    huge={boolean('Huge', false)}
-    giant={boolean('Giant', false)}
     modifier={select('Modifier', modifierOptions, 'hydrogen')}
     variant={select('Variant', variantOptions, 'andromeda')}
     fluid={boolean('Fluid', false)}
     disabled={boolean('Disabled', false)}
     borderRadius={number('Border radius', 0, radiusBorderRange)}
     icon
+    nano={boolean('Nano', false)}
+    micro={boolean('Micro', false)}
+    tiny={boolean('Tiny', false)}
+    big={boolean('Big', false)}
+    huge={boolean('Huge', false)}
+    giant={boolean('Giant', false)}
   >
     <HeartIcon
       width={number('Icon size', 15, svgSizeRange)}
@@ -114,15 +132,16 @@ export const WithTextAndIcon = () => {
   const iconPosition = radios('Icon position', svgPositionOptions, 'left')
   return (
     <Button
-      tiny={boolean('Tiny', false)}
-      big={boolean('Big', false)}
-      huge={boolean('Huge', false)}
-      giant={boolean('Giant', false)}
+      size={select('Size', sizeOptions, 'normal')}
       fluid={boolean('Fluid', false)}
       modifier={select('Modifier', modifierOptions, 'hydrogen')}
       variant={select('Variant', variantOptions, 'andromeda')}
       disabled={boolean('Disabled', false)}
       borderRadius={number('Border radius', 0, radiusBorderRange)}
+      tiny={boolean('Tiny', false)}
+      big={boolean('Big', false)}
+      huge={boolean('Huge', false)}
+      giant={boolean('Giant', false)}
     >
       {iconPosition === 'left' && (
         <HeartIcon
@@ -158,13 +177,14 @@ WithTextAndIcon.decorators = [
 
 export const WithBadge = () => (
   <Button
+    size={select('Size', sizeOptions, 'normal')}
+    modifier={select('Modifier', modifierOptions, 'helium')}
+    variant={select('Variant', variantOptions, 'andromeda')}
+    disabled={boolean('Disabled', false)}
     tiny={boolean('Tiny', false)}
     big={boolean('Big', false)}
     huge={boolean('Huge', false)}
     giant={boolean('Giant', false)}
-    modifier={select('Modifier', modifierOptions, 'helium')}
-    variant={select('Variant', variantOptions, 'andromeda')}
-    disabled={boolean('Disabled', false)}
   >
     <span>{text('Text', 'MyButton')}</span>
     <Badge color={text('Badge color', 'red')} Icon={Cart}>